refactor(team): split message-routing into notifier_*.go (C11)
Mechanical file split of the message-routing cluster from
launcher.go (PLAN.md §C11). Three new themed sibling files; bodies
unchanged. The split preserves the layered responsibility:

  loops -> target resolution -> delivery

Where each layer is now in its own file rather than scattered through
the launcher.

New files:

* notifier_loops.go (122 lines, new) — the three notification poll
  loops (notifyAgentsLoop, notifyTaskActionsLoop,
  notifyOfficeChangesLoop), the safeDeliverMessage panic-recover
  wrapper, and recoverPanicTo (the shared panic logger).

* notifier_targets.go (498 lines, new) — target resolution: given a
  channel message or a task action, who should be woken (immediate
  vs delayed)? Hosts notificationTargetsForMessage (the 200+ line
  routing decision tree), taskNotificationTargets, isChannelDM
  (+ Raw), shouldDeliverDelayedTaskNotification, plus the supporting
  free funcs (shouldBackfillTaskOwner, shouldWakeLeadForTaskAction).
  Also moved here: officeChangeTaskNotifications,
  deliverOfficeChangeNotification, and the
  officeChangeTaskNotification + notificationTarget types because
  they're consumed exclusively by the routing layer.

* notifier_delivery.go (212 lines, new) — delivery once targets
  decided: deliverMessageNotification, deliverTaskNotification,
  taskForAction, taskNotificationContent, sendTaskUpdate,
  sendChannelUpdate, activeHeadlessSlugs, plus the notifyCtx() thin
  wrappers (buildNotificationContext, ultimateThreadRoot,
  threadMessageIDs, buildTaskNotificationContext,
  relevantTaskForTarget, responseInstructionForTarget,
  buildMessageWorkPacket, buildTaskExecutionPacket).

launcher.go: 2670 -> 1883 lines (-787, -29.5% in this PR alone).

Cumulative since main: 4998 -> 1883 = -3115 lines (-62.3%).

Imports trimmed:
  launcher.go drops "runtime"
  notifier_loops.go imports its own minimal set (fmt/os/path/filepath/runtime/strings/time)
  notifier_targets.go imports strings + time
  notifier_delivery.go imports fmt + strings + time

No new tests — pure file split. Existing test suite (32 packages
green) is the safety net. Each function appears in exactly one file
before, exactly one file after, with the same body.

Stacked on refactor/launcher-c10-headless-dispatcher (C10).
