+++ internal/team/launcher.go
@@ -18,7 +18,6 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"runtime"
 	"sort"
 	"strings"
 	"sync"
@@ -482,110 +481,6 @@ func (l *Launcher) Launch() error {
 	return nil
 }
 
-// notifyAgentsLoop subscribes to broker messages and pushes notifications immediately.
-func (l *Launcher) notifyAgentsLoop() {
-	if l.broker == nil {
-		return
-	}
-	msgs, unsubscribe := l.broker.SubscribeMessages(128)
-	defer unsubscribe()
-
-	for msg := range msgs {
-		if l.broker.HasPendingInterview() {
-			continue
-		}
-		if msg.From == "system" {
-			continue
-		}
-		l.safeDeliverMessage(msg)
-	}
-}
-
-// safeDeliverMessage wraps deliverMessageNotification in a panic recover so a
-// bad message doesn't take the whole broker down. Stack is written to stderr
-// and logs/panics.log so we can diagnose the next occurrence.
-func (l *Launcher) safeDeliverMessage(msg channelMessage) {
-	defer recoverPanicTo("deliverMessageNotification", fmt.Sprintf("msg=%+v", msg))
-	l.deliverMessageNotification(msg)
-}
-
-// recoverPanicTo is the shared panic-recovery body used by broker background
-// goroutines. It logs the goroutine stack to stderr and to
-// ~/.wuphf/logs/panics.log so the broker stays up even if a specific action
-// path blows up. Call as: defer recoverPanicTo("loopName", "extra context").
-func recoverPanicTo(site, extra string) {
-	r := recover()
-	if r == nil {
-		return
-	}
-	buf := make([]byte, 16<<10)
-	n := runtime.Stack(buf, false)
-	fmt.Fprintf(os.Stderr, "panic in %s: %v\n%s\n%s\n", site, r, extra, buf[:n])
-	if home, err := os.UserHomeDir(); err == nil {
-		if f, ferr := os.OpenFile(filepath.Join(home, ".wuphf", "logs", "panics.log"), os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0o644); ferr == nil {
-			_, _ = fmt.Fprintf(f, "%s panic in %s: %v\n%s\n%s\n\n", time.Now().UTC().Format(time.RFC3339), site, r, extra, buf[:n])
-			_ = f.Close()
-		}
-	}
-}
-
-func (l *Launcher) notifyTaskActionsLoop() {
-	if l.broker == nil {
-		return
-	}
-	actions, unsubscribe := l.broker.SubscribeActions(128)
-	defer unsubscribe()
-
-	for action := range actions {
-		if l.broker.HasPendingInterview() {
-			continue
-		}
-		if action.Kind != "task_created" && action.Kind != "task_updated" && action.Kind != "task_unblocked" {
-			continue
-		}
-		task, ok := l.taskForAction(action)
-		if !ok {
-			continue
-		}
-		// Skip "done" tasks for task_created / task_updated — the agent that completed
-		// the task should send a follow-up broadcast which wakes CEO via the message
-		// loop. But for task_unblocked the task status is still "in_progress" (it was
-		// just unblocked), so we must never skip it regardless of status.
-		if action.Kind != "task_unblocked" && strings.EqualFold(strings.TrimSpace(task.Status), "done") {
-			continue
-		}
-		func() {
-			defer recoverPanicTo("deliverTaskNotification", fmt.Sprintf("action=%+v task=%+v", action, task))
-			l.deliverTaskNotification(action, task)
-		}()
-	}
-}
-
-func (l *Launcher) notifyOfficeChangesLoop() {
-	if l.broker == nil {
-		return
-	}
-	changes, unsubscribe := l.broker.SubscribeOfficeChanges(128)
-	defer unsubscribe()
-
-	for evt := range changes {
-		// office_reseeded fires after onboarding rewrites the whole roster
-		// (blueprint selection). The interactive claude panes were spawned
-		// from the earlier default team and now point at slugs that are no
-		// longer registered agents — messages typed into them go into a
-		// dead shell. Respawn them against the new roster, outside the
-		// interview guard so it can't be blocked by a half-complete wizard.
-		if evt.Kind == "office_reseeded" {
-			l.respawnPanesAfterReseed()
-			continue
-		}
-		if l.broker.HasPendingInterview() {
-			continue
-		}
-		l.deliverOfficeChangeNotification(evt)
-	}
-}
-
 // respawnPanesAfterReseed restarts the interactive agent panes so they match
 // the newly-seeded roster from onboarding. Best-effort: the codex runtime has
 // no interactive panes, and reconfigureVisibleAgents handles an uninitialised
@@ -614,604 +509,10 @@ func (l *Launcher) respawnPanesAfterReseed() {
 	}
 }
 
-type officeChangeTaskNotification struct {
-	Target  notificationTarget
-	Action  officeActionLog
-	Task    teamTask
-	Content string
-}
-
-func (l *Launcher) deliverOfficeChangeNotification(evt officeChangeEvent) {
-	for _, notification := range l.officeChangeTaskNotifications(evt) {
-		l.sendTaskUpdate(notification.Target, notification.Action, notification.Task, notification.Content)
-	}
-}
-
-func (l *Launcher) officeChangeTaskNotifications(evt officeChangeEvent) []officeChangeTaskNotification {
-	if l == nil || l.broker == nil {
-		return nil
-	}
-
-	kind := strings.TrimSpace(evt.Kind)
-	slug := normalizeChannelSlug(evt.Slug)
-	switch kind {
-	case "member_created", "channel_created", "channel_updated":
-	default:
-		return nil
-	}
-
-	targetMap := l.targeter().PaneTargets()
-	if len(targetMap) == 0 {
-		return nil
-	}
-
-	now := time.Now().UTC().Format(time.RFC3339)
-	seen := make(map[string]struct{})
-	var out []officeChangeTaskNotification
-	for _, task := range l.broker.AllTasks() {
-		owner := strings.TrimSpace(task.Owner)
-		if owner == "" {
-			continue
-		}
-		if !shouldBackfillTaskOwner(kind, slug, task) {
-			continue
-		}
-		enabled := false
-		for _, member := range l.broker.EnabledMembers(task.Channel) {
-			if member == owner {
-				enabled = true
-				break
-			}
-		}
-		if !enabled {
-			continue
-		}
-		target, ok := targetMap[owner]
-		if !ok {
-			continue
-		}
-		key := owner + ":" + task.ID
-		if _, exists := seen[key]; exists {
-			continue
-		}
-		seen[key] = struct{}{}
-		action := officeActionLog{
-			Kind:      "task_updated",
-			Source:    "office",
-			Channel:   normalizeChannelSlug(task.Channel),
-			Actor:     "system",
-			RelatedID: task.ID,
-			CreatedAt: now,
-		}
-		out = append(out, officeChangeTaskNotification{
-			Target:  target,
-			Action:  action,
-			Task:    task,
-			Content: l.taskNotificationContent(action, task),
-		})
-	}
-	return out
-}
-
-func shouldBackfillTaskOwner(kind, slug string, task teamTask) bool {
-	status := strings.ToLower(strings.TrimSpace(task.Status))
-	if status == "done" || status == "canceled" || status == "cancelled" || status == "review" {
-		return false
-	}
-	if task.Blocked {
-		return false
-	}
-	switch kind {
-	case "member_created":
-		return strings.TrimSpace(task.Owner) == slug
-	case "channel_created", "channel_updated":
-		return normalizeChannelSlug(task.Channel) == slug
-	default:
-		return false
-	}
-}
-
 const (
 	agentNotifyCooldown      = 1 * time.Second
 	agentNotifyCooldownAgent = 2 * time.Second
 )
-
-func (l *Launcher) deliverMessageNotification(msg channelMessage) {
-	// demo_seed messages exist purely to make #general feel staffed on first
-	// paint; they must never wake an agent or burn an LLM call. Filter at
-	// the central delivery point (not just notifyAgentsLoop) so other
-	// callers — primeVisibleAgents, replays, future routes — can't bypass
-	// it. Today these don't actually route demo_seed targets because the
-	// lead is the From and Tagged is empty, but a future @all-default
-	// change would silently turn the demo seed into an LLM-burning
-	// broadcast. One filter, one place.
-	if msg.Kind == "demo_seed" {
-		return
-	}
-	immediate, delayed := l.notificationTargetsForMessage(msg)
-
-	// Debounce: use shorter cooldown for human/CEO messages, longer for agent-originated
-	// to prevent agent-to-agent feedback loops (devil's advocate finding #3)
-	isHumanOrCEO := msg.From == "you" || msg.From == "human" || msg.From == "nex" || msg.From == l.targeter().LeadSlug()
-	cooldown := agentNotifyCooldownAgent
-	if isHumanOrCEO {
-		cooldown = agentNotifyCooldown
-	}
-	now := time.Now()
-	filtered := make([]notificationTarget, 0, len(immediate))
-	l.notifyMu.Lock()
-	if l.notifyLastDelivered == nil {
-		l.notifyLastDelivered = make(map[string]time.Time)
-	}
-	for _, t := range immediate {
-		if last, ok := l.notifyLastDelivered[t.Slug]; ok && now.Sub(last) < cooldown {
-			continue
-		}
-		l.notifyLastDelivered[t.Slug] = now
-		filtered = append(filtered, t)
-	}
-	l.notifyMu.Unlock()
-	immediate = filtered
-
-	// Mark implicit public-channel routing targets as active so the UI can show
-	// the ephemeral "X is thinking..." indicator. DMs suppress this signal.
-	isDM, _ := l.isChannelDM(normalizeChannelSlug(msg.Channel))
-	if l.broker != nil && len(immediate) > 0 && (msg.From == "you" || msg.From == "human") && !l.isOneOnOne() && !isDM && len(msg.Tagged) == 0 {
-		slugs := make([]string, 0, len(immediate))
-		for _, t := range immediate {
-			slugs = append(slugs, t.Slug)
-		}
-		l.broker.MarkRoutingTargets(slugs)
-	}
-
-	for _, target := range immediate {
-		l.sendChannelUpdate(target, msg)
-	}
-	// Note: delayed is always empty for message notifications — notificationTargetsForMessage
-	// only ever populates immediate. The delayed path is used only for task notifications
-	// via taskNotificationTargets/deliverTaskNotification.
-	_ = delayed
-}
-
-func (l *Launcher) deliverTaskNotification(action officeActionLog, task teamTask) {
-	immediate, delayed := l.taskNotificationTargets(action, task)
-	if len(immediate) == 0 && len(delayed) == 0 {
-		return
-	}
-	content := l.taskNotificationContent(action, task)
-	for _, target := range immediate {
-		l.sendTaskUpdate(target, action, task, content)
-	}
-	for _, target := range delayed {
-		go func(target notificationTarget, action officeActionLog, task teamTask) {
-			time.Sleep(ceoHeadStartDelay)
-			if !l.shouldDeliverDelayedTaskNotification(target.Slug, action, task) {
-				return
-			}
-			l.sendTaskUpdate(target, action, task, content)
-		}(target, action, task)
-	}
-}
-
-type notificationTarget struct {
-	PaneTarget string
-	Slug       string
-}
-
-func (l *Launcher) taskNotificationTargets(action officeActionLog, task teamTask) (immediate []notificationTarget, delayed []notificationTarget) {
-	targetMap := l.targeter().NotificationTargets()
-	if len(targetMap) == 0 {
-		return nil, nil
-	}
-	lead := l.targeter().LeadSlug()
-	enabledMembers := map[string]struct{}{}
-	disabledMembers := map[string]struct{}{}
-	if l.broker != nil {
-		for _, member := range l.broker.EnabledMembers(task.Channel) {
-			enabledMembers[member] = struct{}{}
-		}
-		for _, member := range l.broker.DisabledMembers(task.Channel) {
-			disabledMembers[member] = struct{}{}
-		}
-	}
-	// Task ownership is an explicit human/CEO assignment. The same bypass that
-	// lets an @-tag wake a wizard-hired specialist applies here: the owner may
-	// have been hired post-seed and not yet in ch.Members. Disabled (muted)
-	// members are still excluded — muting is an explicit silence.
-	actor := strings.TrimSpace(action.Actor)
-	owner := strings.TrimSpace(task.Owner)
-	isAssigned := func(slug string) bool {
-		return slug != "" && (slug == owner || slug == actor)
-	}
-	addImmediate := func(slug string) {
-		if slug == "" {
-			return
-		}
-		if _, muted := disabledMembers[slug]; muted {
-			return
-		}
-		if !isAssigned(slug) && len(enabledMembers) > 0 {
-			if _, ok := enabledMembers[slug]; !ok {
-				return
-			}
-		}
-		if target, ok := targetMap[slug]; ok {
-			immediate = append(immediate, target)
-			delete(targetMap, slug)
-		}
-	}
-	addDelayed := func(slug string) {
-		if slug == "" {
-			return
-		}
-		if _, muted := disabledMembers[slug]; muted {
-			return
-		}
-		if !isAssigned(slug) && len(enabledMembers) > 0 {
-			if _, ok := enabledMembers[slug]; !ok {
-				return
-			}
-		}
-		if target, ok := targetMap[slug]; ok {
-			delayed = append(delayed, target)
-			delete(targetMap, slug)
-		}
-	}
-
-	if owner == "" {
-		if lead != "" && lead != actor {
-			addImmediate(lead)
-		}
-		return immediate, delayed
-	}
-
-	if owner == lead {
-		if lead != "" && lead != actor {
-			addImmediate(lead)
-		}
-		return immediate, delayed
-	}
-
-	// Assigned owners should start immediately when new work lands, especially
-	// for CEO-created or automation-created tasks. This is the bridge between
-	// "policy created work" and "the specialist actually begins moving."
-	//
-	// Exception: do not wake the owner when the task is blocked (unresolved
-	// dependencies). They have no work to do until the blocker clears. They
-	// will be notified via a task_unblocked action when deps resolve.
-	if (action.Kind == "task_created" || action.Kind == "watchdog_alert" || action.Kind == "task_unblocked") && owner != actor && !task.Blocked {
-		addImmediate(owner)
-	} else if owner != actor && action.Kind != "task_created" {
-		addDelayed(owner)
-	}
-
-	if lead != "" && lead != owner && lead != actor && !(action.Kind == "task_created" && actor == lead) && shouldWakeLeadForTaskAction(action, task) {
-		addImmediate(lead)
-	}
-
-	return immediate, delayed
-}
-
-func shouldWakeLeadForTaskAction(action officeActionLog, task teamTask) bool {
-	if strings.TrimSpace(action.Kind) != "task_updated" {
-		return true
-	}
-	actor := strings.TrimSpace(action.Actor)
-	owner := strings.TrimSpace(task.Owner)
-	if actor == "" || owner == "" || actor != owner {
-		return true
-	}
-	if task.Blocked {
-		return true
-	}
-	status := strings.ToLower(strings.TrimSpace(task.Status))
-	review := strings.ToLower(strings.TrimSpace(task.ReviewState))
-	if status == "review" || status == "done" || status == "blocked" {
-		return true
-	}
-	if review == "ready_for_review" || review == "approved" {
-		return true
-	}
-	return false
-}
-
-func (l *Launcher) taskForAction(action officeActionLog) (teamTask, bool) {
-	if l.broker == nil || strings.TrimSpace(action.RelatedID) == "" {
-		return teamTask{}, false
-	}
-	id := strings.TrimSpace(action.RelatedID)
-	for _, task := range l.broker.AllTasks() {
-		if task.ID == id {
-			return task, true
-		}
-	}
-	return teamTask{}, false
-}
-
-func (l *Launcher) shouldDeliverDelayedTaskNotification(targetSlug string, action officeActionLog, task teamTask) bool {
-	if l.broker == nil {
-		return true
-	}
-	if strings.EqualFold(strings.TrimSpace(task.Status), "done") {
-		return false
-	}
-	current, ok := l.taskForAction(action)
-	if !ok {
-		return false
-	}
-	if strings.EqualFold(strings.TrimSpace(current.Status), "done") {
-		return false
-	}
-	if strings.TrimSpace(current.Owner) != "" && strings.TrimSpace(current.Owner) != targetSlug && targetSlug != l.targeter().LeadSlug() {
-		return false
-	}
-	if strings.TrimSpace(current.Owner) == "" && targetSlug != l.targeter().LeadSlug() {
-		return false
-	}
-	return true
-}
-
-// taskNotificationContent delegates to the notificationContextBuilder
-// (PLAN.md §C3). See notification_context.go for the formatting body.
-func (l *Launcher) taskNotificationContent(action officeActionLog, task teamTask) string {
-	return l.notifyCtx().TaskNotificationContent(action, task)
-}
-
-func (l *Launcher) sendTaskUpdate(target notificationTarget, action officeActionLog, task teamTask, content string) {
-	channel := normalizeChannelSlug(task.Channel)
-	if channel == "" {
-		channel = "general"
-	}
-	notification := l.buildTaskExecutionPacket(target.Slug, action, task, content)
-	if l.targeter().ShouldUseHeadlessForTarget(target) {
-		l.enqueueHeadlessCodexTurn(target.Slug, headlessSandboxNote()+notification, channel)
-		return
-	}
-	l.queuePaneNotification(target.Slug, target.PaneTarget, notification)
-}
-
-// isChannelDM returns true if the channel is a DM (either old dm-* format or new Store type).
-// agentTarget returns the agent slug that should receive the DM notification (non-human side).
-// isChannelDM is the public entry point used by dispatch code; targeter
-// reads the same logic via the isChannelDMRaw callback.
-func (l *Launcher) isChannelDM(channelSlug string) (isDM bool, agentTarget string) {
-	return l.isChannelDMRaw(channelSlug)
-}
-
-// isChannelDMRaw resolves whether a channel is a direct-message channel
-// and, if so, which agent it targets. Two formats supported: the legacy
-// "dm-{agent}" slug and the new store format where channel.type == "D".
-func (l *Launcher) isChannelDMRaw(channelSlug string) (isDM bool, agentTarget string) {
-	if IsDMSlug(channelSlug) {
-		return true, DMTargetAgent(channelSlug)
-	}
-	if l.broker != nil {
-		cs := l.broker.ChannelStore()
-		if cs != nil && cs.IsDirectMessageBySlug(channelSlug) {
-			ch, ok := cs.GetBySlug(channelSlug)
-			if ok {
-				members := cs.Members(ch.ID)
-				for _, m := range members {
-					if m.Slug != "human" && m.Slug != "you" {
-						return true, m.Slug
-					}
-				}
-			}
-		}
-	}
-	return false, ""
-}
-
-func (l *Launcher) notificationTargetsForMessage(msg channelMessage) (immediate []notificationTarget, delayed []notificationTarget) {
-	targetMap := l.targeter().NotificationTargets()
-	if len(targetMap) == 0 {
-		return nil, nil
-	}
-	// DMs are isolated: only the target agent gets notified, never CEO or others.
-	if ch := normalizeChannelSlug(msg.Channel); IsDMSlug(ch) {
-		agentSlug := DMTargetAgent(ch)
-		if agentSlug == msg.From {
-			return nil, nil // agent's own message, don't echo back
-		}
-		if target, ok := targetMap[agentSlug]; ok {
-			return []notificationTarget{target}, nil
-		}
-		return nil, nil
-	}
-	// Also check the new Store-based DM format.
-	if ch := normalizeChannelSlug(msg.Channel); !IsDMSlug(ch) {
-		if isDM, agentSlug := l.isChannelDM(ch); isDM {
-			if agentSlug == msg.From {
-				return nil, nil
-			}
-			if target, ok := targetMap[agentSlug]; ok {
-				return []notificationTarget{target}, nil
-			}
-			return nil, nil
-		}
-	}
-	if l.isOneOnOne() {
-		slug := l.oneOnOneAgent()
-		if slug == "" || slug == msg.From {
-			return nil, nil
-		}
-		target, ok := targetMap[slug]
-		if !ok {
-			return nil, nil
-		}
-		return []notificationTarget{target}, nil
-	}
-	lead := l.targeter().LeadSlug()
-	owner := ""
-	if l.broker != nil {
-		owner = l.taskOwnerForMessage(msg)
-	}
-	enabledMembers := map[string]struct{}{}
-	disabledMembers := map[string]struct{}{}
-	if l.broker != nil {
-		for _, member := range l.broker.EnabledMembers(msg.Channel) {
-			enabledMembers[member] = struct{}{}
-		}
-		for _, member := range l.broker.DisabledMembers(msg.Channel) {
-			disabledMembers[member] = struct{}{}
-		}
-	}
-
-	// isExplicit checks whether a slug was explicitly @-tagged by the sender.
-	// Explicit tags bypass the enabledMembers filter so a newly hired specialist
-	// not yet in ch.Members can still be reached. They do NOT bypass ch.Disabled:
-	// an explicit disable is the user's intent to silence the agent, and an
-	// @-tag must not override it.
-	isExplicit := func(slug string) bool { return containsSlug(msg.Tagged, slug) }
-
-	addImmediate := func(slug string) {
-		if slug == "" || slug == msg.From {
-			return
-		}
-		if _, muted := disabledMembers[slug]; muted {
-			return
-		}
-		if !isExplicit(slug) && len(enabledMembers) > 0 {
-			if _, ok := enabledMembers[slug]; !ok {
-				return
-			}
-		}
-		if target, ok := targetMap[slug]; ok {
-			immediate = append(immediate, target)
-			delete(targetMap, slug)
-		}
-	}
-	allowTarget := func(slug string) bool {
-		if slug == "" || slug == msg.From {
-			return false
-		}
-		if _, muted := disabledMembers[slug]; muted {
-			return false
-		}
-		explicit := isExplicit(slug)
-		if !explicit && len(enabledMembers) > 0 {
-			if _, ok := enabledMembers[slug]; !ok {
-				return false
-			}
-		}
-		if slug == lead {
-			return true
-		}
-		// Explicit @-tag: always allow regardless of domain. Domain inference is
-		// for implicit routing only — it should never suppress an explicit mention.
-		if explicit {
-			return true
-		}
-		if owner != "" {
-			return slug == owner
-		}
-		if strings.TrimSpace(msg.Content) == "" && strings.TrimSpace(msg.Title) == "" {
-			return false
-		}
-		return l.messageTargetsAgent(msg, slug)
-	}
-
-	// Focus mode (delegation): CEO routes all work. Specialists only wake
-	// when explicitly tagged by CEO or human. No cross-agent chatter.
-	if l.isFocusModeEnabled() {
-		switch {
-		case msg.From == "you" || msg.From == "human" || msg.Kind == "automation" || msg.From == "nex":
-			// When the human explicitly @tags one or more specialists, deliver directly
-			// to those specialists only. CEO does not need to re-route explicit assignments —
-			// the specialist is already awake and acting. CEO only sees untagged human messages
-			// (general questions, requests that need routing decisions).
-			humanExplicitlyTaggedSpecialists := false
-			for _, slug := range msg.Tagged {
-				if slug == "" || slug == msg.From || slug == lead {
-					continue
-				}
-				// Respect explicit disables. A muted specialist stays muted
-				// even when @-tagged — muting is the user's explicit intent.
-				if _, muted := disabledMembers[slug]; muted {
-					continue
-				}
-				// Explicit @-tag trumps channel-membership. The specialist
-				// may have been hired after #general was seeded and not yet
-				// added to ch.Members; dropping the notification here would
-				// silently re-route the human's direct address to CEO.
-				if target, ok := targetMap[slug]; ok {
-					immediate = append(immediate, target)
-					delete(targetMap, slug)
-					humanExplicitlyTaggedSpecialists = true
-				}
-			}
-			if !humanExplicitlyTaggedSpecialists {
-				// No specialist tagged — CEO decides who handles this.
-				addImmediate(lead)
-			}
-		case msg.From == lead:
-			for _, slug := range msg.Tagged {
-				if slug != lead && allowTarget(slug) {
-					addImmediate(slug)
-				}
-			}
-		default:
-			// Specialist message: wake CEO only if it is a substantive update (not a status ping).
-			// [STATUS] lines are internal progress markers — CEO does not need to re-route on them.
-			isStatusOnly := strings.HasPrefix(strings.TrimSpace(msg.Content), "[STATUS]")
-			if !isStatusOnly {
-				addImmediate(lead)
-			}
-		}
-		return immediate, delayed
-	}
-
-	// Collaborative mode: all agents can see domain-relevant messages
-	switch {
-	case msg.From == "you" || msg.From == "human" || msg.Kind == "automation" || msg.From == "nex":
-		// @all: notify every agent immediately.
-		if containsSlug(msg.Tagged, "all") {
-			addImmediate(lead)
-			for slug := range targetMap {
-				addImmediate(slug)
-			}
-			break
-		}
-		addImmediate(lead)
-		if owner != "" && owner != lead && allowTarget(owner) {
-			addImmediate(owner)
-		}
-		for _, slug := range msg.Tagged {
-			if allowTarget(slug) {
-				addImmediate(slug)
-			}
-		}
-	case msg.From == lead:
-		for _, slug := range msg.Tagged {
-			if allowTarget(slug) {
-				addImmediate(slug)
-			}
-		}
-	case containsSlug(msg.Tagged, lead):
-		addImmediate(lead)
-		if owner != "" && owner != lead && allowTarget(owner) {
-			addImmediate(owner)
-		}
-		for _, slug := range msg.Tagged {
-			if allowTarget(slug) {
-				addImmediate(slug)
-			}
-		}
-	default:
-		// Specialist-to-channel message in collaborative mode: CEO stays in the loop
-		// plus any tagged agents and the task owner.
-		addImmediate(lead)
-		if owner != "" && owner != lead && allowTarget(owner) {
-			addImmediate(owner)
-		}
-		for _, slug := range msg.Tagged {
-			if allowTarget(slug) {
-				addImmediate(slug)
-			}
-		}
-	}
-	return immediate, delayed
-}
 
 func (l *Launcher) watchChannelPaneLoop(channelCmd string) {
 	unhealthyCount := 0
@@ -1746,98 +1047,10 @@ func (l *Launcher) notifyCtx() *notificationContextBuilder {
 	return l.notify
 }
 
-// activeHeadlessSlugs returns the slugs that have non-empty headless
-// queues or active turns at the moment of the call. Locks headlessMu so
-// the snapshot is consistent. The except parameter is the slug being
-// notified — the lead must not list itself as "already active".
-func (l *Launcher) activeHeadlessSlugs(except string) map[string]struct{} {
-	if l == nil {
-		return nil
-	}
-	l.headless.mu.Lock()
-	defer l.headless.mu.Unlock()
-	out := map[string]struct{}{}
-	for workerSlug, queue := range l.headless.queues {
-		if workerSlug == except {
-			continue
-		}
-		if len(queue) > 0 {
-			out[workerSlug] = struct{}{}
-		}
-	}
-	for workerSlug, active := range l.headless.active {
-		if workerSlug == except {
-			continue
-		}
-		if active != nil {
-			out[workerSlug] = struct{}{}
-		}
-	}
-	return out
-}
-
 // Notification-context methods on Launcher are thin wrappers; the bodies
 // live in notification_context.go on notificationContextBuilder. Wrappers
 // kept (rather than removed) so the ~50 in-package callers don't need a
 // rename sweep in this PR — that's a follow-up.
-
-func (l *Launcher) buildNotificationContext(channelSlug, triggerMsgID, threadRootID string, limit int) string {
-	return l.notifyCtx().NotificationContext(channelSlug, triggerMsgID, threadRootID, limit)
-}
-
-func (l *Launcher) ultimateThreadRoot(channelSlug, startID string) string {
-	return l.notifyCtx().UltimateThreadRoot(channelSlug, startID)
-}
-
-func (l *Launcher) threadMessageIDs(channelSlug, rootID string) map[string]struct{} {
-	return l.notifyCtx().ThreadMessageIDs(channelSlug, rootID)
-}
-
-func (l *Launcher) buildTaskNotificationContext(channelSlug, slug string, limit int) string {
-	return l.notifyCtx().TaskNotificationContext(channelSlug, slug, limit)
-}
-
-func (l *Launcher) relevantTaskForTarget(msg channelMessage, slug string) (teamTask, bool) {
-	return l.notifyCtx().RelevantTaskForTarget(msg, slug)
-}
-
-func (l *Launcher) responseInstructionForTarget(msg channelMessage, slug string) string {
-	return l.notifyCtx().ResponseInstructionForTarget(msg, slug)
-}
-
-func (l *Launcher) buildMessageWorkPacket(msg channelMessage, slug string) string {
-	return l.notifyCtx().BuildMessageWorkPacket(msg, slug)
-}
-
-func (l *Launcher) buildTaskExecutionPacket(slug string, action officeActionLog, task teamTask, content string) string {
-	return l.notifyCtx().BuildTaskExecutionPacket(slug, action, task, content)
-}
-
-func (l *Launcher) sendChannelUpdate(target notificationTarget, msg channelMessage) {
-	channel := normalizeChannelSlug(msg.Channel)
-	if channel == "" {
-		channel = "general"
-	}
-	notification := ""
-	if l.isOneOnOne() {
-		notification = fmt.Sprintf(
-			"[New from @%s]: %s\n%s Reply using team_broadcast with my_slug \"%s\" and channel \"%s\" reply_to_id \"%s\". Once you have posted the needed reply, STOP and wait for the next pushed notification.",
-			msg.From, truncate(msg.Content, 1000), l.responseInstructionForTarget(msg, target.Slug), target.Slug, channel, msg.ID,
-		)
-	} else {
-		packet := l.buildMessageWorkPacket(msg, target.Slug)
-		notification = fmt.Sprintf(
-			"%s\n---\n[New from @%s]: %s\n%s This packet is your complete context — do NOT call team_poll or team_tasks. Just do the work and reply via team_broadcast with my_slug \"%s\", channel \"%s\", reply_to_id \"%s\". Once you have posted the needed update, STOP and wait for the next pushed notification.",
-			packet, msg.From, truncate(msg.Content, 1000), l.responseInstructionForTarget(msg, target.Slug), target.Slug, channel, msg.ID,
-		)
-	}
-
-	if l.targeter().ShouldUseHeadlessForTarget(target) {
-		l.enqueueHeadlessCodexTurn(target.Slug, headlessSandboxNote()+notification, channel)
-		return
-	}
-	l.queuePaneNotification(target.Slug, target.PaneTarget, notification)
-}
 
 // shouldUseHeadlessDispatch is a thin wrapper around the targeter; see
 // officeTargeter.ShouldUseHeadless for semantics.
